Refactor: moved helper functions to utils, removed unused .py, and changed default value in config
