+++ config/convert_kognic_annotation_to_t4_sample.yaml
@@ -5,7 +5,7 @@ conversion:
   output_base: ./data/non_annotated_t4_format
   # Directory holding the downloaded OpenLABEL JSON(s) (output of download_kognic_annotation).
   # Files are matched to scenes by filename stem and by OpenLABEL metadata (dataset_id / scene_uuid).
-  annotation_base: ./data/kognic_annotations/semantic_segmentation_poc_tier_iv
+  annotation_base: ./data/kognic_annotations
   # Must match the flag used at download time: true → cuboids in ISO8855 frame, false → Kognic internal frame.
   iso_rotated_cuboids: false
   # Optional: rename Kognic object types to T4 category names, e.g. {car: vehicle.car}.
