+++ perception_dataset/kognic/utils/ego_pose.py
@@ -4,10 +4,11 @@ from typing import Dict, List, Optional
 
 import kognic.io.model as KognicModel
 import numpy as np
-from scipy.spatial.transform import Rotation
+from pyquaternion import Quaternion
 
 from perception_dataset.constants import LIDAR_CONCAT_CHANNEL
 from perception_dataset.utils.logger import configure_logger
+from perception_dataset.utils.transform import matrix_to_quaternion_wxyz, transform_matrix
 
 logger = configure_logger(modname=__name__)
 
@@ -41,15 +42,14 @@ def extract_ego_poses(
         logger.warning("No ego poses found")
         return {}
 
-    t0 = _build_transform(ego_pose_by_token[frame_ego_tokens[0]])
-    t0_inv = np.linalg.inv(t0)
+    t0_inv = _build_transform(ego_pose_by_token[frame_ego_tokens[0]], inverse=True)
 
     ego_poses: Dict[str, KognicModel.EgoVehiclePose] = {}
     for frame_idx, ego_token in enumerate(frame_ego_tokens):
         transform = _build_transform(ego_pose_by_token[ego_token])
         relative_transform = t0_inv @ transform
         position = relative_transform[:3, 3]
-        quat = _matrix_to_quat_wxyz(relative_transform[:3, :3])
+        quat = matrix_to_quaternion_wxyz(relative_transform[:3, :3])
 
         ego_poses[str(frame_idx)] = KognicModel.EgoVehiclePose(
             position=KognicModel.Position(
@@ -91,70 +91,10 @@ def _get_reference_sample_data(
     return next(iter(frame_record.values()), None)
 
 
-def _build_transform(ego_pose: dict) -> np.ndarray:
-    transform = np.eye(4)
-    rotation = ego_pose["rotation"]
-    if Rotation is not None:
-        transform[:3, :3] = Rotation.from_quat(
-            [rotation[1], rotation[2], rotation[3], rotation[0]]
-        ).as_matrix()
-    else:
-        transform[:3, :3] = _quat_wxyz_to_matrix(rotation)
-    transform[:3, 3] = ego_pose["translation"]
-    return transform
-
-
-def _quat_wxyz_to_matrix(quat: list) -> np.ndarray:
-    w, x, y, z = [float(v) for v in quat]
-    norm = w * w + x * x + y * y + z * z
-    scale = 2.0 / norm
-
-    wx, wy, wz = scale * w * x, scale * w * y, scale * w * z
-    xx, xy, xz = scale * x * x, scale * x * y, scale * x * z
-    yy, yz, zz = scale * y * y, scale * y * z, scale * z * z
-
-    return np.array(
-        [
-            [1.0 - (yy + zz), xy - wz, xz + wy],
-            [xy + wz, 1.0 - (xx + zz), yz - wx],
-            [xz - wy, yz + wx, 1.0 - (xx + yy)],
-        ],
-        dtype=np.float64,
+def _build_transform(ego_pose: dict, inverse: bool = False) -> np.ndarray:
+    """Build a 4x4 transform from a T4 ego_pose record (rotation is ``[w, x, y, z]``)."""
+    return transform_matrix(
+        translation=np.array(ego_pose["translation"]),
+        rotation=Quaternion(ego_pose["rotation"]),
+        inverse=inverse,
     )
-
-
-def _matrix_to_quat_wxyz(matrix: np.ndarray) -> np.ndarray:
-    if Rotation is not None:
-        quat_xyzw = Rotation.from_matrix(matrix).as_quat()
-        return np.array(
-            [quat_xyzw[3], quat_xyzw[0], quat_xyzw[1], quat_xyzw[2]],
-            dtype=np.float64,
-        )
-
-    trace = float(np.trace(matrix))
-    if trace > 0.0:
-        scale = np.sqrt(trace + 1.0) * 2.0
-        w = 0.25 * scale
-        x = (matrix[2, 1] - matrix[1, 2]) / scale
-        y = (matrix[0, 2] - matrix[2, 0]) / scale
-        z = (matrix[1, 0] - matrix[0, 1]) / scale
-    elif matrix[0, 0] > matrix[1, 1] and matrix[0, 0] > matrix[2, 2]:
-        scale = np.sqrt(1.0 + matrix[0, 0] - matrix[1, 1] - matrix[2, 2]) * 2.0
-        w = (matrix[2, 1] - matrix[1, 2]) / scale
-        x = 0.25 * scale
-        y = (matrix[0, 1] + matrix[1, 0]) / scale
-        z = (matrix[0, 2] + matrix[2, 0]) / scale
-    elif matrix[1, 1] > matrix[2, 2]:
-        scale = np.sqrt(1.0 + matrix[1, 1] - matrix[0, 0] - matrix[2, 2]) * 2.0
-        w = (matrix[0, 2] - matrix[2, 0]) / scale
-        x = (matrix[0, 1] + matrix[1, 0]) / scale
-        y = 0.25 * scale
-        z = (matrix[1, 2] + matrix[2, 1]) / scale
-    else:
-        scale = np.sqrt(1.0 + matrix[2, 2] - matrix[0, 0] - matrix[1, 1]) * 2.0
-        w = (matrix[1, 0] - matrix[0, 1]) / scale
-        x = (matrix[0, 2] + matrix[2, 0]) / scale
-        y = (matrix[1, 2] + matrix[2, 1]) / scale
-        z = 0.25 * scale
-
-    return np.array([w, x, y, z], dtype=np.float64)
